Add a selection tool icon
git-svn-id: svn://svn.code.sf.net/p/qtvlm/code/branches/qtVlm_dev@1624 2387b160-4621-4e8a-be36-40ea10d72da2

+++ src/Grib.h
@@ -89,7 +89,7 @@ class Grib
 
         std::set<time_t>  * getListDates()   {return &setAllDates;}
 
-        int        getNumberOfDates()      {return setAllDates.size();}
+        int        getNumberOfDates()      {return (int)setAllDates.size();}
         time_t     getRefDate()            {return setAllDates.empty() ?
                                                        0 : *setAllDates.begin();}
         time_t     getMinDate()            {return ok?minDate:-1; }
